+++ NEWS.md
@@ -1,7 +1,7 @@
 # store 0.4.0
 
 
-* Added **as_gridref**, **precision** S3 class of grid reference functions using parent 
+* Added **as_gridref**, **precision**, **projection** S3 class of grid reference functions using parent 
 **gridCoords** function from [archived rnbn package](https://github.com/ropensci-archive/rnbn/issues/37).
 S3 classes described in [S3](https://adv-r.hadley.nz/s3.html) chapter of Advanced R
 * Added **grid_reference_hectad**, **grid_reference_pentad**, **grid_reference_tetrad**,
@@ -9,7 +9,7 @@ S3 classes described in [S3](https://adv-r.hadley.nz/s3.html) chapter of Advance
 to convert higher resolution grid references using parent **gridRef** function from [archived rnbn package](https://github.com/ropensci-archive/rnbn/issues/37)
 * Added **extract_paragraph_delimited_list** to family of extract_delimited_list
 functions. This function adds two carriage returns and new lines as a delimiter
-* Added **grid_reference_projection**, **grid_reference_easting**, **grid_reference_northing** 
+* Added **grid_reference_easting**, **grid_reference_northing** 
 family of grid reference functions using parent **gridCoords** function from [archived rnbn package](https://github.com/ropensci-archive/rnbn/issues/37)
 * Added **vc_grid_square_intersects** A data frame listing the vice counties intersects
 with 10km, 2km and 1km grid squares